runner/requirements.txt
bidict==0.22.0
blessed==1.19.1
certifi==2022.6.15.2
charset-normalizer==2.1.1
codefind==0.1.3
dnspython==2.2.1
eventlet==0.33.1
frida==15.2.2
greenlet==1.1.3
idna==3.4
jurigged==0.5.2
ovld==0.3.2
psutil==5.9.2
py-cpuinfo==8.0.0
python-engineio==4.3.4
python-socketio==5.7.1
requests==2.28.1
six==1.16.0
urllib3==1.26.12
watchdog==2.1.9
wcwidth==0.2.5
websocket-client==1.4.1
websockets==10.3
